BERT: Deprecate XLA_DOWNCAST_BF16 and use model.to(torch.bfloat16)

+++ torch-neuronx/training/dp_bert_hf_pretrain/adamw_fp32_optim_params.py
@@ -98,7 +98,7 @@ class AdamW_FP32OptimParams(Optimizer):
                 if self.upcast_optim_states:
                     grad = p.grad.data.double()
                 else:
-                    grad = p.grad.data
+                    grad = p.grad.data.float()
 
                 if grad.is_sparse:
                     raise RuntimeError("Adam does not support sparse gradients, please consider SparseAdam instead")
